requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.40
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.9
aioredis==2.0.1
fastapi-cache2==0.2.1
structlog==24.1.0 
prometheus-client==0.20.0
python-dotenv==1.0.1
PyJWT==2.8.0 
email-validator==2.2.0
dnspython==2.7.0
bcrypt==4.3.0
alembic==1.13.1 